strain display created
